handle refactor of admin contoller
