+++ src/main/java/com/example/contributoryloanapp/service/serviceImplementation/AdminService.java
@@ -1,5 +1,7 @@
 package com.example.contributoryloanapp.service.serviceImplementation;
 
+import com.example.contributoryloanapp.payload.request.LoginRequest;
+import com.example.contributoryloanapp.payload.response.JwtResponse;
 import com.example.contributoryloanapp.payload.response.auth.ForgotPasswordResponse;
 import com.example.contributoryloanapp.payload.response.auth.ResetPassword;
 import org.springframework.http.ResponseEntity;
