+++ src/main/java/com/example/contributoryloanapp/service/serviceImplementation/UserServiceImpl.java
@@ -8,13 +8,17 @@ import com.example.contributoryloanapp.mapper.UserMapper;
 import com.example.contributoryloanapp.model.ERole;
 import com.example.contributoryloanapp.model.Role;
 import com.example.contributoryloanapp.model.User;
+import com.example.contributoryloanapp.payload.request.LoginRequest;
 import com.example.contributoryloanapp.payload.request.SignupRequest;
+import com.example.contributoryloanapp.payload.response.JwtResponse;
 import com.example.contributoryloanapp.payload.response.auth.ForgotPasswordResponse;
 import com.example.contributoryloanapp.payload.response.auth.EditUser;
 import com.example.contributoryloanapp.payload.response.auth.ResetPassword;
 import com.example.contributoryloanapp.payload.response.auth.UpdatePasswordRequest;
 import com.example.contributoryloanapp.repository.RoleRepository;
 import com.example.contributoryloanapp.repository.UserRepository;
+import com.example.contributoryloanapp.security.jwt.JwtUtils;
+import com.example.contributoryloanapp.security.services.UserDetailsImpl;
 import com.example.contributoryloanapp.service.UserService;
 import com.example.contributoryloanapp.utils.DateUtils;
 import com.example.contributoryloanapp.utils.classes.RoleAssignment;
@@ -25,6 +29,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.BadCredentialsException;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -50,10 +58,6 @@ public class UserServiceImpl implements UserService {
     @Autowired
     private BCryptPasswordEncoder bCryptPasswordEncoder;
 
-    @Bean
-    public BCryptPasswordEncoder bCryptPasswordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
 
     @Autowired
     private RoleAssignment roleAssignment;
@@ -64,6 +68,12 @@ public class UserServiceImpl implements UserService {
 
     @Autowired
     private ModelMapper modelMapper;
+
+    @Autowired
+    AuthenticationManager authenticationManager;
+
+    @Autowired
+    JwtUtils jwtUtils;
 
 
     @Override
@@ -95,6 +105,59 @@ public class UserServiceImpl implements UserService {
 
         return user;
     }
+
+    public JwtResponse authenticateUser(LoginRequest loginRequest){
+
+        Authentication authentication = null;
+        try {
+            authentication = authenticationManager.authenticate(new
+                    UsernamePasswordAuthenticationToken(
+                    loginRequest.getUsername(),
+                    loginRequest.getPassword()));
+
+        } catch (BadCredentialsException e) {
+
+            throw new ApiRequestException("Incorrect Username or Password");
+
+        }
+
+        //context holder holding the authentication object for future validation and authorization
+        SecurityContextHolder.getContext().setAuthentication(authentication);
+
+        String jwt = jwtUtils.generateJwtToken(authentication);
+
+        UserDetailsImpl userDetails = (UserDetailsImpl) authentication.getPrincipal();
+
+        // UserDetails userDetails1 =  userDetailsService.loadUserByUsername(loginRequest.getUsername());
+        //alternative to the line above to retrieve userdetails
+
+        List<String> roles = userDetails.getAuthorities().stream()
+                .map(item->item.getAuthority())
+                .collect(Collectors.toList());
+
+        return  new JwtResponse(jwt,
+                userDetails.getId(),
+                userDetails.getUsername(),
+                userDetails.getEmail(),
+                roles);
+
+    }
+
+    public ResponseEntity<String> updatePassword(UpdatePasswordRequest updatePasswordRequest){
+
+        User user = getLoggedInUser();
+
+        checkIfValidOldPassword(user, updatePasswordRequest);
+
+        boolean  isPassWordChanged = changeUserPassword( user,   updatePasswordRequest);
+
+        if(isPassWordChanged){
+
+            return ResponseEntity.ok("Password Changed Successfully");
+        }
+        return new ResponseEntity<>("Operation failed!!", HttpStatus.BAD_REQUEST);
+    }
+
 
     private boolean isValidPassword(String password) {
         String regex = "^(([0-9]|[a-z]|[A-Z]|[@])*){8,20}$";
